Take menuButtons into account for the caption's availableRect.
It will still center the caption if there's room for the menu to not
cover it. The caption will be pushed right if there's no room in the
center. The "graident masking" effect is visible as we hadn't pushed
it far enough away, but I think I prefer it this way as it's a good
visualization that this text is not a menu item.