Log instance id to confirm invisible button overtop "File".
By logging the instance id we confirm that the bug is caused by the
buttons not properly being removed from Decoration::Private::buttons
when we call ButtonGroup::removeButton(). It only erases the value
from ButtonGroup::Private::buttons, but doesn't appear to remove it
from the decoration's list.

Because it's not removed from the decoration's list, the button still
recieves mouse events, and recieves the button click first.

The button is only removed from the decoration's list when the button
is destroyed, as there's a connection to QObject::destroyed made when
Decoration::addButton is called.