Experimental support for dragging AppMenuButton to move (Issue #2)
DecorationButton::isPressed() remains true however, even after
releasing the mouse. Not only is there a visual bug, but due to how
KDecoration mouse events work, this means that clicking an AppMenu
button after the currently "pressed" button will not open it's menu on
the first click. Instead it will "reset" the pressed state of the
previously pressed button.