Move drag logic to Decoration.
We only start the drag if the presss is within the menuButtons rect.

We use the pressed hack on all menuButtons if a dragStarted.

There's still a bug when we start a drag a few pixels from the top of
the titlebar, then hoverLeave dragging upwards. The dragged point
resets to (0,0) but the drag is started, causing it the mouse to jump
to the titlebars (0,0) and switch to the move cursor.