Pass wheelEvents, middle+right clicks to decoration (Issue #2)
We can also setAccepted(false) of LeftButton. Unfortunately I don't
think there's a way to detect when the titlebar starts the drag, so
we can't run unPressAllButtons to fix the sticky button underneath the
drag. It causes the (possibly) wrong menu to open on next click.

So we're stuck with manually sending a move event to X11.