Add option to change the titlebar opacity (Issue #8)
Use QFormLayout. Not sure if passing telling KCModule to watch itself
is a good idea in `addConfig(skel, this)`. Might need to wrap the
form layout with a VBox + QWidget like Aurorae does. Rewriting to use
a Qt UI file would make this irrelevant anyways.

Moved shadowColor to Material.h and call updateShadow on reconfigure.