First crack at showing menu on hover (Issue #16)
I first tried only displaying when the menu itself is hovered but it
didn't feel right. The user has no way of knowing how wide the menu is
so I'm going to display the menu if anywhere in the title is hovered.

For some reason, Decoration::hoverEnterEvent+hoverLeaveEvent don't
appear to be called.

Since the enter/leave properties seem to be weird, I'll probably
refactor to change state on Decoration::sectionUnderMouseChanged.